# Removes fs2.internal.Future and inlines the implementation in to Task