… core/jvm/src/main/scala/fs2/TaskPlatform.scala
… core/shared/src/main/scala/fs2/internal/Future.scala