Exploration of inlining Future in to Task. `Task#map` is optimized in this version.

Can't merge this to 0.9 because it is not a binary compatible change. If we are interested in merging this, it'll have to wait for 0.10.