@mpilquist 2016-10-01
Some benchmarks:

`jmh:run -i 10 -wi 10 -f1 -t1 --jvmArgs -Xms2g --jvmArgs -Xmx2g .*TaskBenchmark.*`

```
## Baseline

[info] Benchmark                             Mode  Cnt       Score      Error  Units
[info] TaskBenchmark.apply_12800            thrpt   10   42578.243 ± 3449.131  ops/s
[info] TaskBenchmark.current_12800          thrpt   10  226894.359 ± 9619.846  ops/s
[info] TaskBenchmark.flatMap_12800          thrpt   10    1087.893 ±  128.304  ops/s
[info] TaskBenchmark.map_12800              thrpt   10     688.280 ±  227.691  ops/s
[info] TaskBenchmark.parallelTraverse_1600  thrpt   10     172.695 ±   35.158  ops/s
[info] TaskBenchmark.race_12800             thrpt   10   19917.140 ± 1759.230  ops/s
[info] TaskBenchmark.start_12800            thrpt   10   20454.268 ± 1518.413  ops/s
[info] TaskBenchmark.traverse_1600          thrpt   10    3905.010 ±  267.189  ops/s

## Combined Task+Future

[info] Benchmark                             Mode  Cnt       Score      Error  Units
[info] TaskBenchmark.apply_12800            thrpt   10   31165.628 ± 2516.559  ops/s
[info] TaskBenchmark.current_12800          thrpt   10  210209.289 ± 7836.701  ops/s
[info] TaskBenchmark.flatMap_12800          thrpt   10    1156.195 ±   71.813  ops/s
[info] TaskBenchmark.map_12800              thrpt   10     927.821 ±   25.556  ops/s
[info] TaskBenchmark.parallelTraverse_1600  thrpt   10     178.820 ±   18.059  ops/s
[info] TaskBenchmark.race_12800             thrpt   10   20776.244 ± 2609.083  ops/s
[info] TaskBenchmark.start_12800            thrpt   10   20280.987 ± 2368.089  ops/s
[info] TaskBenchmark.traverse_1600          thrpt   10    4887.936 ±  162.897  ops/s

## Combined Task+Future + optimized map (what's in this PR)

[info] Benchmark                             Mode  Cnt       Score       Error  Units
[info] TaskBenchmark.apply_12800            thrpt   10   31328.833 ±  1805.603  ops/s
[info] TaskBenchmark.current_12800          thrpt   10  212880.162 ± 13285.017  ops/s
[info] TaskBenchmark.flatMap_12800          thrpt   10    1055.998 ±   148.690  ops/s
[info] TaskBenchmark.map_12800              thrpt   10    6105.499 ±  1453.048  ops/s
[info] TaskBenchmark.parallelTraverse_1600  thrpt   10     174.867 ±    25.236  ops/s
[info] TaskBenchmark.race_12800             thrpt   10   20517.391 ±  3043.132  ops/s
[info] TaskBenchmark.start_12800            thrpt   10   16604.427 ±  3719.253  ops/s
[info] TaskBenchmark.traverse_1600          thrpt   10    4849.137 ±   183.422  ops/s
```