@pchlupacek 2016-10-03
@mpilquist cool. Just some more thoughts
 - Task is trampoline + async + fail. Cannot we somehow take advantage of this and reduce boxing even further where we use Task and Trampoline together? 
-  If I read performance results correctly, apply is slower so as start. Which basically shows the only thing that really takes advantage is the optimized map. Perhaps we could get map in 0.9 and see if we cannot do better with breaking changes?