@pchiusano 2016-10-03
I'd be in favor of this if only because it simplifies the code.

An idea I had is that rather than `Trampoline[Unit]`, you might be able to get away with something like `List[() => Unit]` (or `Catenable[() => Unit]`). Not sure if this will be more efficient, but if you aren't using any intermediate values other than `Unit` inside the trampoline, you don't need the trampoline type's extra structure.