@mpilquist 2016-11-12
Running:
```
jmh:run -i 10 -wi 10 -f1 -t1 --jvmArgs -Xms2g --jvmArgs -Xmx2g .*TaskBenchmark.apply
```

Before:
```
[info] Benchmark                   Mode  Cnt      Score       Error  Units
[info] TaskBenchmark.apply_1      thrpt   10  67205.231 ±  3519.402  ops/s
[info] TaskBenchmark.apply_4      thrpt   10  70051.840 ±  4391.744  ops/s
[info] TaskBenchmark.apply_100    thrpt   10  70992.649 ±  3485.477  ops/s
[info] TaskBenchmark.apply_200    thrpt   10  70012.790 ±  6383.783  ops/s
[info] TaskBenchmark.apply_400    thrpt   10  66782.414 ±  4104.506  ops/s
[info] TaskBenchmark.apply_800    thrpt   10  67337.552 ±  2801.884  ops/s
[info] TaskBenchmark.apply_1600   thrpt   10  56217.385 ±  4751.706  ops/s
[info] TaskBenchmark.apply_3200   thrpt   10  61195.078 ±  3096.898  ops/s
[info] TaskBenchmark.apply_6400   thrpt   10  53361.838 ±  5886.136  ops/s
[info] TaskBenchmark.apply_12800  thrpt   10  42551.507 ± 10475.376  ops/s
```

After:
```
[info] Benchmark                   Mode  Cnt      Score      Error  Units
[info] TaskBenchmark.apply_1      thrpt   10  69822.907 ± 4160.554  ops/s
[info] TaskBenchmark.apply_4      thrpt   10  68595.642 ± 5185.729  ops/s
[info] TaskBenchmark.apply_100    thrpt   10  69168.573 ± 3143.320  ops/s
[info] TaskBenchmark.apply_200    thrpt   10  68160.382 ± 5151.096  ops/s
[info] TaskBenchmark.apply_400    thrpt   10  65383.883 ± 3693.868  ops/s
[info] TaskBenchmark.apply_800    thrpt   10  66677.737 ± 3053.816  ops/s
[info] TaskBenchmark.apply_1600   thrpt   10  63798.693 ± 3677.837  ops/s
[info] TaskBenchmark.apply_3200   thrpt   10  58021.485 ± 4257.027  ops/s
[info] TaskBenchmark.apply_6400   thrpt   10  54225.153 ± 3334.463  ops/s
[info] TaskBenchmark.apply_12800  thrpt   10  33395.930 ± 3661.071  ops/s
```