@mpilquist 2016-11-12
series/1.0

```
[info] Benchmark                   Mode  Cnt      Score      Error  Units
[info] TaskBenchmark.start_1      thrpt   10  32646.956 ± 2876.661  ops/s
[info] TaskBenchmark.start_4      thrpt   10  33973.265 ± 2864.659  ops/s
[info] TaskBenchmark.start_100    thrpt   10  37439.694 ± 4501.797  ops/s
[info] TaskBenchmark.start_200    thrpt   10  31787.168 ± 5628.920  ops/s
[info] TaskBenchmark.start_400    thrpt   10  35335.519 ± 5864.458  ops/s
[info] TaskBenchmark.start_800    thrpt   10  35978.453 ± 4012.692  ops/s
[info] TaskBenchmark.start_1600   thrpt   10  29703.707 ± 3354.544  ops/s
[info] TaskBenchmark.start_3200   thrpt   10  26195.198 ± 2073.370  ops/s
[info] TaskBenchmark.start_6400   thrpt   10  24493.295 ± 1696.017  ops/s
[info] TaskBenchmark.start_12800  thrpt   10  20293.333 ± 1249.530  ops/s
```

This branch:

```
[info] Benchmark                   Mode  Cnt      Score      Error  Units
[info] TaskBenchmark.start_1      thrpt   10  35447.752 ± 2957.790  ops/s
[info] TaskBenchmark.start_4      thrpt   10  34555.406 ±  857.775  ops/s
[info] TaskBenchmark.start_100    thrpt   10  34317.816 ±  831.255  ops/s
[info] TaskBenchmark.start_200    thrpt   10  33514.636 ± 2841.553  ops/s
[info] TaskBenchmark.start_400    thrpt   10  32577.232 ± 2882.115  ops/s
[info] TaskBenchmark.start_800    thrpt   10  37379.757 ± 1892.682  ops/s
[info] TaskBenchmark.start_1600   thrpt   10  37033.801 ± 5153.598  ops/s
[info] TaskBenchmark.start_3200   thrpt   10  26703.492 ± 3109.320  ops/s
[info] TaskBenchmark.start_6400   thrpt   10  26045.342 ± 1419.290  ops/s
[info] TaskBenchmark.start_12800  thrpt   10  20745.177 ±  506.656  ops/s
```